Add businessName, receiverPhone. Code cleanups
